// app/model/campaign.revenue.model.tsx
/// <reference path="../../typings/index.d.ts" />
import * as React from 'react';

export class RevenueAll {
    RevenueID: number = 0;
    ExitTrafficCampaignID: string;
    PeriodStart: Date;
    PeriodEnd: Date;
    PayoutType: string;
    PayoutRate: string;
    BillableCount: string;
    Revenue: string;
    RevenueCurrency: string;
    DisableAutomaticUpdate: string;
    UpdatedDateTime: Date;
    ConfirmedForInvoicing: string;
    Invoiced: string;
    InvoicedDate: string;
    InvoiceAmount: string;
    InvoiceCurrency: string;
    NotInvoicedReason: string;
    InvoiceNotes: string;    
}